Backend/requirements.txt
alembic==1.10.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cors==1.0.1
decorator==5.1.1
distlib==0.3.6
filelock==3.11.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
numpy==1.24.2
packaging==23.1
pipenv==2023.3.20
platformdirs==3.2.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==1.4.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
requests-file==1.5.1
six==1.16.0
spoonacular==3.0
SQLAlchemy==2.0.9
timedelta==2020.12.3
tldextract==3.4.0
typing_extensions==4.5.0
urllib3==1.26.15
validators==0.20.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
Werkzeug==2.2.3
WTForms==3.0.1
zope.event==4.6
zope.interface==6.0
